requirements.txt
altair==5.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
beautifulsoup4==4.12.3
deep-translator==1.11.4
diffusers==0.30.3
gdown==5.2.0
google-generativeai==0.7.1
Jinja2==3.1.3
jupyterlab==4.2.2
keras==3.4.0
librosa==0.10.2
Markdown==3.6
matplotlib==3.9.0
notebook==7.2.1
numpy==1.26.4
opencv-python==4.8.0.76
pandas==2.2.2
pillow==10.3.0
pint==0.24.2
plotly==5.23.0
PyAutoGUI==0.9.54
pymongo==4.8.0
PyPDF2==3.0.1
pyperclip==1.9.0
python-barcode==0.15.1
pyttsx3==2.90
pywhatkit==5.4
pywikihow==0.5.7
qrcode==7.4.2
requests==2.31.0
setuptools==70.1.1
SpeechRecognition==3.10.3
spotipy==2.24.0
speedtest-cli==2.1.3
streamlit==1.38.0
streamlit-ws-localstorage==1.0.6
tensorflow==2.16.1
torch==2.3.0
transformers==4.45.2
ultralytics==8.3.3
websockets==13.0.1
